requirements.txt
flask
flask-restful
flask-cors
google-generativeai
python-dotenv
pymysql